pyproject.toml
[tool.poetry]
name = "financialtransactionswindget"
version = "0.1.0"
description = ""
authors = ["const <leonovkw@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
logging = "^0.4.9.6"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"
types-requests = "^2.32.0.20241016"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'


[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.git          # исключить .git в корне проекта
  )/
)
'''


[tool.isort]
# максимальная длина строки
line_length = 119
